mingdeng/requirements.txt
# 核心依赖
streamlit          # Web 界面框架
mem0ai           # 记忆系统（可选）
openai            # OpenAI 兼容 API 客户端
python-dateutil>=2.8.0     # 日期处理

# UI 增强
streamlit-aggrid   # 表格组件（支持拖拽）
plotly            # 可视化图表
pandas             # 数据处理

# 工具库
pydantic           # 数据验证
